+++ Control/Joint/Base/Maybe.hs
@@ -4,15 +4,6 @@ import Control.Joint.Composition (Composition (Primary, unwrap))
 import Control.Joint.Transformer (Transformer (Schema, lay, wrap))
 import Control.Joint.Schemes.UT (UT (UT))
 
-instance Composition Maybe where
-	type Primary Maybe a = Maybe a
-	unwrap x = x
-
-instance Transformer Maybe where
-	type Schema Maybe u = UT Maybe u
-	lay x = UT $ Just <$> x
-	wrap x = UT . pure $ x
-
 instance Functor u => Functor (UT Maybe u) where
 	fmap f (UT x) = UT $ (fmap . fmap) f x
 
@@ -22,3 +13,12 @@ instance Applicative u => Applicative (UT Maybe u) where
 
 instance (Applicative u, Monad u) => Monad (UT Maybe u) where
 	UT x >>= f = UT $ x >>= maybe (pure Nothing) (unwrap . f)
+
+instance Composition Maybe where
+	type Primary Maybe a = Maybe a
+	unwrap x = x
+
+instance Transformer Maybe where
+	type Schema Maybe u = UT Maybe u
+	lay x = UT $ Just <$> x
+	wrap x = UT . pure $ x