+++ Control/Joint/Base/Reader.hs
@@ -6,15 +6,6 @@ import Control.Joint.Schemes.TU (TU (TU))
 
 newtype Reader e a = Reader (e -> a)
 
-instance Composition (Reader e) where
-	type Primary (Reader e) a = (->) e a
-	unwrap (Reader x) = x
-
-instance Transformer (Reader e) where
-	type Schema (Reader e) u = TU ((->) e) u
-	lay x = TU . const $ x
-	wrap x = TU $ pure <$> unwrap x
-
 instance Functor u => Functor (TU ((->) e) u) where
 	fmap f (TU x) = TU $ \r -> f <$> x r
 --
@@ -24,3 +15,12 @@ instance Applicative u => Applicative (TU ((->) e) u) where
 
 instance (Applicative u, Monad u) => Monad (TU ((->) e) u) where
 	TU x >>= f = TU $ \e -> x e >>= ($ e) . unwrap . f
+
+instance Composition (Reader e) where
+	type Primary (Reader e) a = (->) e a
+	unwrap (Reader x) = x
+
+instance Transformer (Reader e) where
+	type Schema (Reader e) u = TU ((->) e) u
+	lay x = TU . const $ x
+	wrap x = TU $ pure <$> unwrap x