+++ Control/Joint/Schemes.hs
@@ -1,4 +1,5 @@
 module Control.Joint.Schemes (module Exports) where
 
-import Control.Joint.Schemes.UTU as Exports
+import Control.Joint.Schemes.TUT as Exports
+import Control.Joint.Schemes.UT as Exports
 import Control.Joint.Schemes.TU as Exports