+++ Control/Joint/Base/Reader.hs
@@ -1,7 +1,7 @@
 module Control.Joint.Base.Reader where
 
 import Control.Joint.Composition (Composition (Primary, run))
-import Control.Joint.Transformer (Transformer (Schema, embed, build))
+import Control.Joint.Transformer (Transformer (Schema, embed, build, unite))
 import Control.Joint.Schemes.TU (TU (TU))
 
 newtype Reader e a = Reader (e -> a)
@@ -24,3 +24,4 @@ instance Transformer (Reader e) where
 	type Schema (Reader e) u = TU ((->) e) u
 	embed x = TU . const $ x
 	build x = TU $ pure <$> run x
+	unite = TU