+++ CHANGELOG.md
@@ -7,3 +7,4 @@
 # 0.1.2
 * Define `Functor`/`Applicative`/`Monad` instances for transformers
 * Define `State` datatype with `get`, `modify` and `put` operations
+* Rename and fix `UTU` joint scheme to `TUT`